requirements/requirements.txt
#python_version >= '3.12'

PySide6
geopandas
openpyxl
pandas
python-docx
selenium
requests
rasterio
owslib
odfdo
bs4
pikepdf

#pip install -r Requirements.txt

#pip freeze > UninstallRequirements.txt
#pip uninstall -r UninstallRequirements.txt -y
